exporters/moonbeam/requirements.txt
pyaml==23.5.9
requests
flask==2.3.2
numpy==1.24.3

